requirements.txt
comtypes==1.4.2
EasyProcess==1.1
entrypoint2==1.1
MouseInfo==0.1.3
mss==9.0.1
numpy==1.24.4
pillow==10.3.0
PyAutoGUI==0.9.54
pygame==2.5.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
pytweening==1.2.0
pywin32==306
pywinauto==0.6.8
rapidfuzz==3.9.0
screen-ocr==0.5.0
six==1.16.0
winsdk==1.0.0b10
